// src/app/clients/open-f1/models/driver/driver.interface.ts
export interface IDriver  {
    broadcastName: string
    countryCode: string
    driverNumber: number
    firstName: string
    fullName: string
    headshotUrl: string
    lastName: string
    meetingKey: number
    nameAcronym: string
    sessionKey: number
    teamColour: string
    teamName: string
}