// src/app/clients/open-f1/models/driver/driver.properties.ts
export enum DriverProperties {
    BROADCAST_NAME = 'broadcast_name',
    COUNTRY_CODE = 'country_code',
    DRIVER_NUMBER = 'driver_number',
    FIRST_NAME = 'first_name',
    FULL_NAME = 'full_name',
    HEADSHOT_URL = 'headshot_url',
    LAST_NAME = 'last_name',
    MEETING_KEY = 'meeting_key',
    NAME_ACRONYM = 'name_acronym',
    SESSION_KEY = 'session_key',
    TEAM_COLOUR = 'team_colour',
    TEAM_NAME = 'team_name',
}